Rich text: extract paste handler
